Исправлены некоторые типы: i64 для timestamp. Исправлены массивы: референсные массивы генерировались как структуры.

+++ alor_client/src/models/bodyrequest_orders_actions_limit.rs
@@ -27,7 +27,7 @@ pub struct BodyrequestOrdersActionsLimit {
   #[serde(rename = "Instrument")]
   instrument: BodyrequestOrdersActionsLimitInstrument, 
   #[serde(rename = "OrderEndUnixTime")]
-  order_end_unix_time: i32,  // 0 
+  order_end_unix_time: i64,  // 0 
   #[serde(rename = "Price")]
   price: Decimal,  // 195.11 
   #[serde(rename = "Quantity")]
@@ -39,7 +39,7 @@ pub struct BodyrequestOrdersActionsLimit {
 }
 
 impl BodyrequestOrdersActionsLimit {
-  pub fn new(instrument: BodyrequestOrdersActionsLimitInstrument, order_end_unix_time: i32, price: Decimal, quantity: i32, side: Operation, user: BodyrequestOrdersActionsLimitUser, ) -> BodyrequestOrdersActionsLimit {
+  pub fn new(instrument: BodyrequestOrdersActionsLimitInstrument, order_end_unix_time: i64, price: Decimal, quantity: i32, side: Operation, user: BodyrequestOrdersActionsLimitUser, ) -> BodyrequestOrdersActionsLimit {
     BodyrequestOrdersActionsLimit {
       instrument: instrument,
       order_end_unix_time: order_end_unix_time,
@@ -64,16 +64,16 @@ impl BodyrequestOrdersActionsLimit {
   }
 
 
-  pub fn set_order_end_unix_time(&mut self, order_end_unix_time: i32) {
+  pub fn set_order_end_unix_time(&mut self, order_end_unix_time: i64) {
     self.order_end_unix_time = order_end_unix_time;
   }
 
-  pub fn with_order_end_unix_time(mut self, order_end_unix_time: i32) -> BodyrequestOrdersActionsLimit {
+  pub fn with_order_end_unix_time(mut self, order_end_unix_time: i64) -> BodyrequestOrdersActionsLimit {
     self.order_end_unix_time = order_end_unix_time;
     self
   }
 
-  pub fn order_end_unix_time(&self) -> &i32 {
+  pub fn order_end_unix_time(&self) -> &i64 {
     &self.order_end_unix_time
   }
